-Sleep recorder sleeptime double fix -Sleep recorder toasts -Sleep review fixed small sleep time -Fixed sleep calendar date and hour format -Fixed float->double in Sleepdata

+++ app/src/main/java/com2027/cw/group7/resteasy/SleepData.java
@@ -24,7 +24,7 @@ public class SleepData extends FireStore {
     }
 
     public SleepData(long userRating, long sleepRating,
-                      float userSleepTime, String treatment, String comment, String date) {
+                      double userSleepTime, String treatment, String comment, String date) {
         this.userRating = userRating;
         this.sleepRating = sleepRating;
         this.userSleepTime = userSleepTime;
